news_bot/requirements.txt
requests
beautifulsoup4
feedparser
python-dotenv
google-generativeai
fastapi
uvicorn
pydantic
